refactor(plugin-axios,plugin-fetch): share clientGenerator via internals/client
plugin-axios's and plugin-fetch's clientGenerator.tsx were identical
106-line files except for the plugin type and the generator's own
name ('axios'/'fetch'). Adds createClientGenerator to internals/client,
mirroring the existing createSdkGenerator factory in the same
directory (both plugins already call that one identically). Each
plugin's clientGenerator.tsx now just instantiates the shared factory,
preserving the concrete `clientGenerator` export both packages'
index.ts already expose publicly, so plugin.ts and the test call
sites need no changes at all.

Relocates ContractClientFactory from sdkGenerator.tsx into types.ts so
both generator factories share one definition.

No behavior change, verified via each package's snapshot test suite.
