+++ internals/client/src/index.ts
@@ -13,6 +13,7 @@ export { buildGroupedOptionsSignature } from './builders/signature.ts'
 export { buildValidatorHooks } from './builders/validator.ts'
 export { Operation } from './components/Operation.tsx'
 export { SdkClient } from './components/SdkClient.tsx'
+export { createClientGenerator } from './generators/clientGenerator.tsx'
 export { createSdkGenerator } from './generators/sdkGenerator.tsx'
 export { defaultMacros } from './macros.ts'
 export { resolveClient, resolveContractClient } from './resolveClient.ts'
@@ -20,4 +21,4 @@ export type { ClientSelector, ResolveClientResult, ResolvedContractClient } from
 export { resolveClientOperation } from './resolveClientOperation.ts'
 export type { ClientOperation } from './resolveClientOperation.ts'
 export { resolverClient } from './resolver.ts'
-export type { Mode, Options, ValidatorOptions, ResolvedOptions, ResolverClient } from './types.ts'
+export type { ContractClientFactory, Mode, Options, ValidatorOptions, ResolvedOptions, ResolverClient } from './types.ts'
