+++ internals/client/src/types.ts
@@ -155,3 +155,10 @@ export type ResolvedOptions = {
  * builders, resolver, and components to one option/resolver shape.
  */
 export type PluginContractClient = PluginFactoryOptions<'plugin-contract-client', Options, ResolvedOptions, ResolverClient>
+
+/**
+ * The shape any client plugin (plugin-fetch, plugin-axios) must satisfy to reuse the shared
+ * client and SDK generators. Pins the option, resolved-option, and resolver shapes while leaving
+ * the plugin name free.
+ */
+export type ContractClientFactory = PluginFactoryOptions<string, Options, ResolvedOptions, ResolverClient>
